requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiogram==3.10.0
aiogram-widgets==1.2.7
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.7.4
click==8.1.7
dictdiffer==0.9.0
frozenlist==1.4.1
idna==3.7
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
soupsieve==2.5
StrEnum==0.4.15
tomlkit==0.13.0
tortoise-orm==0.21.5
typing_extensions==4.12.2
yarl==1.9.4
